Move BaseController to actor.py
